Adds a (very rough!) API for LimitlessLED bulbs
This is pretty unpolished.  I wanted to get it committed so that I can
stick to my "every Sunday I must get something in" goal -- otherwise,
I'd probably never get the drive to finish it.  Hopefully now the shame
of an ugly commit will lead me to refactoring it!

In particular, the database handling is rough, as is the type
conversions.  Also, now I have 3 config files -- that seems silly (all
the config for these should be in a database anyway!).